Wrote Test Code for a 'GoToShotAngle' Command with Button as well as deleted Pneumatics Code
