+++ src/org/usfirst/frc4930/Zoot/Robot.java
@@ -29,7 +29,7 @@ public class Robot extends IterativeRobot {
 	public static HookExtender hookExtender;
 	public static IntakeMotors intakeMotors;
 	public static LimitSwitch limitSwitch;
-	public static Pneumatics pneumatics;
+	public static MoveArmToShotAngle moveArmToShotAngle;
 	public static Roller roller;
 
 	public void robotInit() {
@@ -46,7 +46,7 @@ public class Robot extends IterativeRobot {
 		hookExtender = new HookExtender();
 		intakeMotors = new IntakeMotors();
 		limitSwitch = new LimitSwitch();
-		pneumatics = new Pneumatics();
+		moveArmToShotAngle = new MoveArmToShotAngle();
 		roller = new Roller();
 
 		// OI must be instantiated after subsystems
