+++ src/org/usfirst/frc4930/Zoot/subsystems/ArmLifter.java
@@ -10,8 +10,8 @@ import edu.wpi.first.wpilibj.command.Subsystem;
  */
 public class ArmLifter extends Subsystem {
 	
-	private final double VERTICAL_POT_EXTREME = 1.0;
-	private final double HORIZONTAL_POT_EXTREME = 0.0;
+	private final double HIGH_POT_EXTREME = 1.0;
+	private final double LOW_POT_EXTREME = 0.0;
 	
 	public void initDefaultCommand() {
 		setDefaultCommand(new MoveArm());
@@ -19,7 +19,7 @@ public class ArmLifter extends Subsystem {
 
 	public void move(double yAxis) {
 		double armPotVal = RobotMap.armPot.get();
-		if ((armPotVal >= HORIZONTAL_POT_EXTREME) && (armPotVal <= VERTICAL_POT_EXTREME)) {
+		if ((armPotVal >= LOW_POT_EXTREME) && (armPotVal <= HIGH_POT_EXTREME)) {
 			RobotMap.armDart.set(yAxis);
 		}		
 	}
