+++ src/org/usfirst/frc4930/Zoot/subsystems/MoveArmToShotAngle.java
@@ -1,0 +1,14 @@
+package org.usfirst.frc4930.Zoot.subsystems;
+
+import org.usfirst.frc4930.Zoot.commands.GoToShotAngle;
+
+import edu.wpi.first.wpilibj.command.Subsystem;
+
+public class MoveArmToShotAngle extends Subsystem{
+
+	@Override
+	protected void initDefaultCommand() {
+		setDefaultCommand(new GoToShotAngle());
+	}
+
+}
